+++ src/Element/Queries/Concerns/QueriesCustomFields.php
@@ -54,13 +54,6 @@ trait QueriesCustomFields
 
     protected function initQueriesCustomFields(): void
     {
-        foreach ([
-            ...array_keys(Fields::allFieldHandles()),
-            ...array_keys(Fields::allGeneratedFieldHandles()),
-        ] as $handle) {
-            $this->customFieldValues[$handle] = null;
-        }
-
         $this->beforeQuery(function (ElementQuery $elementQuery) {
             // Gather custom fields and generated field handles
             $elementQuery->customFields = [];
@@ -198,20 +191,25 @@ trait QueriesCustomFields
 
         $fieldsByHandle = $this->fieldsByHandle($elementQuery);
 
-        foreach (array_keys(Fields::allFieldHandles()) as $handle) {
+        $fieldHandles = Fields::allFieldHandles();
+
+        foreach ($elementQuery->customFieldValues as $handle => $value) {
             // $fieldAttributes->$handle will return true even if it's set to null, so can't use isset() here
             if ($handle === 'owner') {
                 continue;
             }
-            if (($elementQuery->customFieldValues[$handle] ?? null) === null) {
+            if (! array_key_exists($handle, $fieldHandles)) {
+                continue;
+            }
+            if ($value === null) {
                 continue;
             }
             // Make sure the custom field exists in one of the field layouts
             if (! isset($fieldsByHandle[$handle])) {
                 // If it looks like null/:empty: is a valid option, let it slide
-                $value = is_array($elementQuery->customFieldValues[$handle]) && isset($elementQuery->customFieldValues[$handle]['value'])
-                    ? $elementQuery->customFieldValues[$handle]['value']
-                    : $elementQuery->customFieldValues[$handle];
+                $value = is_array($value) && isset($value['value'])
+                    ? $value['value']
+                    : $value;
 
                 if (is_array($value) && in_array(null, $value, true)) {
                     $values = [...$value];
@@ -224,16 +222,16 @@ trait QueriesCustomFields
                 throw new QueryAbortedException("No custom field with the handle \"$handle\" exists in the field layouts involved with this element query.");
             }
 
-            $glue = $elementQuery->customFieldValues[$handle] === ':empty:'
+            $glue = $value === ':empty:'
                 ? QueryParam::AND
                 : QueryParam::OR;
 
-            $this->where(function (Builder $query) use ($fieldsByHandle, $glue, $handle, $elementQuery) {
+            $this->where(function (Builder $query) use ($fieldsByHandle, $glue, $handle, $value) {
                 foreach ($fieldsByHandle[$handle] as $instances) {
-                    $query->where(function (Builder $query) use ($handle, $elementQuery, $instances) {
+                    $query->where(function (Builder $query) use ($instances, $value) {
                         static::$activeQuery = $this;
                         try {
-                            $instances[0]::modifyQuery($query, $instances, $elementQuery->customFieldValues[$handle]);
+                            $instances[0]::modifyQuery($query, $instances, $value);
                         } finally {
                             static::$activeQuery = null;
                         }
@@ -288,5 +286,18 @@ trait QueriesCustomFields
         }
 
         return $fieldsByHandle;
+    }
+
+    private function isCustomFieldHandle(string $handle): bool
+    {
+        if (array_key_exists($handle, $this->customFieldValues)) {
+            return true;
+        }
+
+        if (array_key_exists($handle, Fields::allFieldHandles())) {
+            return true;
+        }
+
+        return array_key_exists($handle, Fields::allGeneratedFieldHandles());
     }
 }